# docs/swagger.yaml
basePath: /v1
definitions:
...
  models.DeckConfig:
    properties:
      settings_today:
        type: boolean
    type: object
...
  models.LoginResponse:
    properties:
      message:
        type: string
      token:
        type: string
    type: object
  models.LoginStruct:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  models.PasswordResetConfirm:
    properties:
      code:
        type: string
      email:
        type: string
      password:
        type: string
    type: object
  models.RegisterStruct:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  models.ResponseAuth:
    properties:
      message:
        type: string
      success:
        type: boolean
      user:
        $ref: '#/definitions/models.User'
    type: object
...
  models.User:
    properties:
      email:
        type: string
      user_avatar:
        example: avatar url
        type: string
      user_bio:
        example: A simple demo bio
        type: string
      user_name:
        description: This should be unique
        example: Yume
        type: string
      user_permissions:
        description: '0: User; 1: Mod; 2: Admin'
        example: 0
        type: integer
    type: object
host: http://192.168.1.151:1813/
info:
  contact:
    email: contact@memnix.app
    name: API Support
  description: Memnix API
  license:
    name: BSD 3-Clause License
    url: https://github.com/memnix/memnix-rest/blob/main/LICENSE
  termsOfService: https://github.com/memnix/memnix/blob/main/PRIVACY.md
  title: Memnix
  version: "1.0"
paths:
...
  /v1/login:
    post:
      description: Login the user and return a fresh token
      parameters:
      - description: Credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.LoginStruct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Incorrect password or email
        "500":
          description: Internal error
      summary: logins user and return a fresh token
      tags:
      - Auth
  /v1/logout:
    post:
      description: Logout the user and create a record in the log
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "401":
          description: Forbidden
      security:
      - Beaver: []
      summary: logouts the user
      tags:
      - Auth
...
  /v1/register:
    post:
      description: Create a new user
      parameters:
      - description: Credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.RegisterStruct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "403":
          description: Forbidden
      summary: creates a new user
      tags:
      - Auth
  /v1/user:
    get:
      description: Get connected user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAuth'
        "401":
          description: Forbidden
      security:
      - Beaver: []
      summary: gets connected user
      tags:
      - Auth
  /v1/users:
    get:
      deprecated: true
      description: Get all users.  Shouldn't really be used
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - Admin: []
      summary: gets a list of user
      tags:
      - User
  /v1/users/confirmpassword:
    post:
      consumes:
      - application/json
      description: Confirm a password reset
      parameters:
      - description: Password reset
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/models.PasswordResetConfirm'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: reset a password
      tags:
      - User
  /v1/users/id/{id}:
    get:
      description: Get a user by ID.
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - Admin: []
      summary: gets a user
      tags:
      - User
  /v1/users/resetpassword:
    post:
      consumes:
      - application/json
      description: Request a password reset
      parameters:
      - description: Email
        in: body
        name: config
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: gets a code to reset a password
      tags:
      - User
  /v1/users/settings/{deckId}/today:
    post:
      consumes:
      - application/json
      description: Set the today config for a deck
      parameters:
      - description: Deck ID
        in: path
        name: deckId
        required: true
        type: integer
      - description: Deck Config
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/models.DeckConfig'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: sets the today config for a deck
      tags:
      - User
securityDefinitions:
  Admin:
    in: header
    name: Authorization
    type: apiKey
  Beaver:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"